Flyouts cleanup
