+++ FluentAvalonia/UI/Controls/MenuFlyout/MenuFlyoutSeparator.cs
