+++ FluentAvalonia/UI/Controls/MenuFlyout/MenuFlyoutSubItemTemplate.cs
@@ -2,24 +2,38 @@
 using Avalonia.Controls.Templates;
 using Avalonia.Data;
 using Avalonia.Data.Core;
-using Avalonia.Input;
 using Avalonia.Markup.Parsers;
 using Avalonia.Markup.Xaml.MarkupExtensions;
-using Avalonia.Metadata;
 using System;
 
 namespace FluentAvalonia.UI.Controls
 {
+	/// <summary>
+	/// Represents a <see cref="IDataTemplate"/> that is used to generate items in a 
+	/// <see cref="MenuFlyout"/>
+	/// </summary>
 	public class MenuFlyoutSubItemTemplate : ITreeDataTemplate
 	{
+		/// <summary>
+		/// The type of data being bound
+		/// </summary>
 		public Type DataType { get; set; }
 
+		/// <summary>
+		/// The binding for the <see cref="MenuFlyoutSubItem.Text"/> property
+		/// </summary>
 		[AssignBinding]
 		public BindingBase HeaderText { get; set; }
 		
+		/// <summary>
+		/// The binding for the <see cref="MenuFlyoutSubItem.Icon"/> property
+		/// </summary>
 		[AssignBinding]
 		public BindingBase Icon { get; set; }
 
+		/// <summary>
+		/// The binding to resolve the <see cref="MenuFlyoutSubItem.Items"/>
+		/// </summary>
 		[AssignBinding]
 		public BindingBase SubItems { get; set; }
